Restructured project files
