+++ typescript/packages/group-tree-plot/src/components/PlotErrorOverlay.tsx
