+++ typescript/packages/group-tree-plot/src/components/TreePlotRenderer/index.ts
@@ -1,0 +1,1 @@
+export { TreePlotRenderer } from "./TreePlotRenderer";
