+++ typescript/packages/group-tree-plot/src/utils/treePlot.ts
@@ -1,9 +1,10 @@
-import _ from "lodash";
-import type { DataAssembler } from "./utils/DataAssembler";
-import type { D3TreeEdge, D3TreeNode } from "./types";
-import React from "react";
+/**
+ * Utilities for handling the tree nodes generated by d3
+ */
 
-export const TREE_TRANSITION_DURATION = 200;
+import _ from "lodash";
+import type { D3TreeEdge, D3TreeNode } from "../types";
+import type { DataAssembler } from "./dataAssembler";
 
 export function computeNodeId(node: D3TreeNode) {
     if (!node.parent) {
@@ -12,16 +13,13 @@ export function computeNodeId(node: D3TreeNode) {
         return `${node.parent.data.node_label}_${node.data.node_label}`;
     }
 }
-
 export function computeLinkId(link: D3TreeEdge) {
     return `path ${computeNodeId(link.target)}`;
 }
-
 export function printTreeValue(value: number | null): string {
     if (value === null) return "N/A";
     else return value.toFixed(0);
 }
-
 export function findClosestVisibleInNewTree(
     targetNode: D3TreeNode,
     newTreeRoot: D3TreeNode | null
@@ -58,7 +56,6 @@ export function findClosestVisibleInNewTree(
 
     return visibleParent;
 }
-
 export function findClosestVisibleParent(
     assembler: DataAssembler,
     treeNode: D3TreeNode
@@ -97,7 +94,6 @@ export function findClosestVisibleParent(
     // ? I think this might be unreachable?
     return visibleParent;
 }
-
 export function diagonalPath(link: D3TreeEdge): string {
     const { source, target } = link;
 
@@ -110,13 +106,4 @@ export function diagonalPath(link: D3TreeEdge): string {
             ${avgY} ${target.x}, \
             ${target.y} ${target.x}
         `;
-}
-
-export function usePrevious<T>(value: T): T | null {
-    const oldValRef = React.useRef<T | null>(null);
-    React.useEffect(() => {
-        oldValRef.current = value;
-    }, [value]);
-
-    return oldValRef.current;
 }
